@@ -1,13 +1,8 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import ElementUI from 'element-ui';
-import http from './utils/http'
-import 'element-ui/lib/theme-chalk/index.css';
+import router from './router'
+import store from './store'
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css'
 
-Vue.use(ElementUI);
-Vue.use(http);
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).use(store).use(router).use(ElementPlus).mount('#app')
